+++ types/excel-script.d.ts
@@ -13,6 +13,9 @@ declare namespace ExcelScript {
     getWorksheets(): Worksheet[];
     getWorksheet(name: string): Worksheet | undefined;
     addWorksheet(name?: string): Worksheet;
+    getActiveWorksheet(): Worksheet;
+    getSelectedRange(): Range;
+    getActiveCell(): Range;
   }
 
   interface Worksheet {
@@ -20,8 +23,10 @@ declare namespace ExcelScript {
     getVisibility(): SheetVisibility;
     getUsedRange(): Range | undefined;
     delete(): void;
-    getRange(address: string): Range;
+    getRange(address?: string): Range;
     getRangeByIndexes(startRow: number, startColumn: number, rowCount: number, columnCount: number): Range;
+    setShowGridlines(show: boolean): void;
+    setTabColor(color: string | undefined): void;
   }
 
   interface Range {
@@ -29,15 +34,24 @@ declare namespace ExcelScript {
     getRowCount(): number;
     getColumnCount(): number;
     setValues(values: (string | number | boolean)[][]): void;
+    setNumberFormat(format: string): void;
     getFormat(): RangeFormat;
+    getTables(): Table[];
+  }
+
+  interface Table {
+    getRange(): Range;
+    setPredefinedTableStyle(style: string): void;
   }
 
   interface RangeFormat {
     getFont(): Font;
     autofitColumns(): void;
+    setColumnWidth(width: number): void;
   }
 
   interface Font {
     setBold(value: boolean): void;
+    setSize(size: number): void;
   }
 }